requirements.txt
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
transformers>=4.35.0
diffusers>=0.25.0
accelerate>=0.25.0
peft>=0.7.0
datasets>=2.14.0
safetensors>=0.4.0
huggingface-hub>=0.19.0
tensorboard>=2.14.0
scipy>=1.10.0
pandas>=2.0.0
scikit-image>=0.21.0
lpips>=0.1.4
PyQt5>=5.15.0
